requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml
duckduckgo-search
pypdf
bs4
arxiv
pymupdf
wikipedia
faiss-cpu
langchain_google_genai
langchain_huggingface
langchain-pinecone
pandas
openai
pymongo
pillow
pdfplumber
tqdm
camelot-py[base]==1.0.0
langchain-mongodb
pymongo
marimo[recommended]